SCRUM-43 refactor(Alert): refactoring the Alert component and its dependencies. Specifically, types were move to a separate file. Add mising  documentation and exports
